Make sure a directory still exists on the server before deleting it
The recent changes to WordPress Develop (src) needing it to be built to test it introduced an interesting case where a directory could exists into the database but no more on the server. In this case instead of returning false too early when a user tries to delete it from the UI, we need to let the function carry on so that the directory can be deleted from the database.

+++ inc/media.php
@@ -456,10 +456,6 @@ function mediatheque_delete_dir( $dir = null ) {
 	$dirpath    = get_post_meta( $dir_id, '_mediatheque_relative_path', true );
 	$dirpath    = trailingslashit( $uploadpath['basedir'] ) . $dirpath;
 
-	if ( ! is_dir( $dirpath ) ) {
-		return false;
-	}
-
 	$children = get_children( array(
 		'post_type' => 'user_media',
 		'post_parent' => $dir->ID,
@@ -482,8 +478,10 @@ function mediatheque_delete_dir( $dir = null ) {
 		}
 	}
 
-	// Remove the empty directory
-	@ rmdir( $dirpath );
+	// Remove the directory.
+	if ( is_dir( $dirpath ) ) {
+		@ rmdir( $dirpath );
+	}
 
 	$dir = wp_delete_post( $dir_id, true );
